+++ docs/api/services/aio/agentic_resume.md
@@ -1,0 +1,3 @@
+# Agentic Resume (async)
+
+::: superme_sdk.services.aio._agentic_resume.AsyncAgenticResumeMixin
