+++ docs/api/services/aio/groups.md
@@ -1,0 +1,3 @@
+# Groups (async)
+
+::: superme_sdk.services.aio._groups.AsyncGroupsMixin
