+++ superme_sdk/aio/__init__.py
@@ -1,0 +1,5 @@
+"""Async transport layer for SuperMe SDK."""
+
+from ._http import AsyncHttpMixin
+
+__all__ = ["AsyncHttpMixin"]
